referential_lines/show: Search by name and stop area name
Use Ransack to filter results by stop area name in addition to the
existing filter by name. Xin came up with this Ransack parameter.

We'll need to additionally filter the query so that only the first and
last stop areas are looked at.

There's also a problem where when searching by name, three copies of the
route appear in the listing. Need to investigate what's up with that.

Refs #3126