Chouette::Route: Add class method to include first stop point
We need to be able to get the first and last stop points (including stop
areas, so they can be filtered by name). This is for the search field on
the `ReferentialLinesController#show` page.

Here as a first step, we get just the first stop point. Need to work out
how to include the last stop point and stop areas.

Refs #3126