vehicle_journey_spec.rb: Simplify factory call
Simplify the "orders journeys with nil times at the end" test. Instead
of manually initialising our data, create a special factory for the
empty vehicle journey that we need. This allows us to get rid of the
nonsense in the test setup.

The existing `:vehicle_journey` factory still works the same as before.

Refs #3268