vehicle_journey_spec.rb: #with_stops should sort journeys missing stops
If a journey is missing one or more stops at the beginning of the route,
it should be sorted by the first stop it stops at.

Given these journeys:

           |   1   |   2   |   3
    -------+-------+-------+------
    Stop 1 | 10:00 |       | 10:15
    Stop 2 | 10:10 | 10:12 | 10:25
    Stop 3 | 10:20 | 10:18 | 10:35

Journey should appear in between 1 and 3 because its first stop is
between 10:00 and 10:15.

Move the `initialize_stop_times` helper function outside of the first
test in this `describe` so we can use it again in this test.

Refs #3268