VehicleJourneysController: Move departure time filter to a new method
Like we did in d0544404ffd77d276a5db5e8cacc85e26bc5a272 with
`filter_without_time_tables`, add a similar method to extract the
departure time filter from `#collection`. This allows us to keep the
`#collection` method a little cleaner.

Add `maybe_` prefixes to both aforementioned new filter methods because
we aren't guaranteed to be filtering. It depends on the contents of the
request parameters received.

Remove the comment describing the date range filter now that it has a
method name.

Refs #3427