VehicleJourney: Make `#exclude_journeys_without_time_tables` do opposite
Cédric clarified the task:
https://projects.af83.io/projects/stif-boiv-realisation/activity?from=2017-05-22

Apparently, the "Afficher les courses sans calendrier" filter needs to
only show vehicle journeys that have no calendar associated when toggled
to "Oui".

When I was implementing this, I had thought that it was the other way
around, and the default neutral state was "Oui", and when set to "Non",
vehicle journeys without calendars would be excluded from the list. This
was apparently wrong.

The default state should actually be "Non", and only non-calendared
vehicle journeys should appear when "Oui".

Update the test and code to reflect this. Rename
`#exclude_journeys_without_time_tables` to `#without_time_tables`
because it should now only select vehicle journeys without time tables.

We can also simplify our query quite a bit, since vehicle journeys that
have no time tables will always be distinct in the left join, so we
don't need to explicitly write an extra step in the join to do that.

Refs #3427