VehicleJourney: Move `#increasing_times` validator to a new class
Create a new validator class for this validation. By splitting the
increasing time validation into a separate group, we can unify the
functionality that's currently split between `VehicleJourney` and
`VehicleJourneyAtStop`. This will allow us to go to a single place to
look for the validation.

Move the `#increasing_times` method into our new class with minor
modification to make it work. Move the test for the method into a new
spec file.

Refs #870