Create `TimeDuration` module for `#exceeds_gap?` method
Put the `exceeds_gap?` method in a new module called `TimeDuration`.
This will allow us to use it from both the
`Chouette::VehicleJourneyAtStop` class and from the
`IncreasingTimeOrderValidator`. Doing this also allows us to test it
independently.

Haven't yet rewritten the increasing times validator to take advantage
of this new method, but adding it now to ensure we don't get an error
calling a non-existent method in `VehicleJourneyAtStop`.

I changed the signature of the method to take a duration as its first
argument. This allows us to set arbitrary durations to compare against
and surfaces the duration when reading code where the  method is called.
Not sure if that's the best approach here, but it seems to make sense
for now.

Refs #870