IncreasingTimeOrder: Make `increasing_times_validate` a class method
Making this a class method allows us to test it more easily. Previously
it was attached to `Chouette::VehicleJourneyAtStop`, and the tests I
copied call it on a VehicleJourneyAtStop.

It seems like instantiating a Validator class is kind of a pain, or at
least a bit weird, since it's for internal Rails use. So we change it to
a class method.

Change `#exceeds_gap?` in this class to a class method also so we can
continue to call it from `.increasing_times_validate`. Eventually we'll
switch to the new version in the `TimeDuration` module.

Comment out the tests for `#exceeds_gap?` for now as we'll be moving
those into a new spec file for `TimeDuration`

Refs #870