TimeDuration: Move `.exceeds_gap?` specs into the module's spec file
Take the specs out of the `IncreasingTimeOrderValidator` spec and move
them into `TimeDuration`. This is part of the consolidation and complete
move of the `exceeds_gap?` method into `TimeDuration`.

Add a context to the two test blocks that specifies the duration at a 4
hour gap. Currently the implementation doesn't do anything with that
argument, but this makes it clearer that we're using a four hour
duration.

Refs #870