IncreasingTimeOrderValidator: Move #increasing_times to #validate_each
Get rid of the `#increasing_times` method since it no longer made sense
to me to keep it separate. We're not doing anything extra in
`#validate_each`, the only place where it's called, so move the method's
contents to `#validate_each`.

Rename the `record` argument to `vehicle_journey` as I expect this
validator to only be used on a `VehicleJourney`.

Rename the `describe` in the specs because we're not actually calling
`#increasing_times` in the spec.

Refs #870