Add a basic test to check VehicleJourneyAtStop day offsets
Create a new class that will determine the day offset of a
VehicleJourneyAtStop given a list of VehicleJourneyAtStops.

The idea is, every time midnight is passed, the day offset should be
incremented and saved in the appropriate fields of the
VehicleJourneyAtStop. No real code right now, Just committing the first
test I came up with.

The test takes a handful of at-stops and checks that they have the
correct `departure_day_offset` & `arrival_day_offset` after running our
calculation method.

Refs #3596