vehicle_journey_at_stops_day_offset_spec.rb: Get rid of hard-coded date
I used '2000-01-01' because that's what the VehicleJourneyfactory does
for VehicleJourneyAtStops.

Ultimately, though, explicitly setting a date is confusing because the
`arrival_time` & `departure_time` fields should be time-only, and not
have dates.

The date is added automatically but isn't used and it shouldn't matter
that it's there. (My guess is that it's added by the database adapter or
something like that.)

To make this code less confusing and make it clear that we only care
about the times, get rid of the dates.

Refs #3596