vehicle_journey_at_stops_day_offset_spec.rb: Fill in other 2 test cases
Add the test code for the other two cases we had originally thought of.
Here we test that the offset is correct when both `arrival_time` &
`departure_time` are on the same day, and then that the day offset will
get incremented correctly when there are multiple days represented.

In order for multiple days to work, there need to be intermediary
at-stops that occur on the offsetted day.

Refs #3596