VehicleJourneyAtStopsDayOffset: Restructure methods for persistence
The `#calculate!` method didn't actually save anything to the database.
Deciding to keep the `!` because it does modify the object's internal
at-stops.

I didn't call an update on the at-stop objects in `#calculate!` on
purpose as I wanted to avoid hitting the database in the tests to try to
keep them from getting too much slower than they already are.

But, we need to be able to persist the day_offset values when this class
gets integrated. To do so, adding an `#update` method that will run the
calculation function and update the objects in the database. Split out
the actual updating into a new `#save` method because that seems
logically separate. Decided not to test these two methods as the real
work we care about is in `#calculate!`.

Refs #3596