Add in-progress `TableBuilderHelper` (WIP)
Porting `table_builder` code over from `NewapplicationHelper`. The code
is very much in progress and incomplete. Only decided to commit it now
because I keep making changes and want to have those tracked so I don't
lose anything.

A lot of the original is commented out and the initial test I wrote
doesn't pass. The test also doesn't use the proper data, I just copied
it from the website's output.

Anyway, this is sort of the starting point.

Refs #3479