TableBuilder: Port `links_builder` and `sortable_columns` from old code
Bring these two methods from `NewapplicationHelper` into
`TableBuilderHelper` so we can mess with them in the context of our
helper. We'll want to try to change & improve them a bit. Also need to
make sure we rename them, otherwise they'll conflict with the existing
helper methods.

Added a few notes to myself in the comments.

Refs #3479