TableBuilder: Ideas for method arguments
A couple of ideas for new method arguments. Didn't like `cls:`, but not
a fan of avoiding the name just for looks if the alternative is messy.

Don't like the fact that `current_referential` comes in as a global to a
couple parts of the builder. Ideally want to be explicit about that it
is and where it comes from. Unfortunately, it's rather convoluted and
dynamic, designed to abstract away the difference between different
models' referentials. And it's connected to `parent` (very dynamic) via
the `inherited_resources` gem, which means that refactoring it out would
create a huge mess of changes. Not anxious to board that ship just yet.

Do we want to pass `current_referential` in directly and just make the
tests easier to write? Would it be a pain to pass it in every time from
the caller? I don't know yet. Still trying to figure out the best course
of action there.

Refs #3479