TableBuilder spec: Stub a `params` global
Since the `TableBuilder` depends on the controller's `params` hash (the
link builder needs the controller, action, etc. in order to build a
link), we need to stub `params` in order for the helper call to work in
our test.

Still not a fan of this implicit dependency, but we'll see. I might keep
it in because maybe it makes more sense and would require less change to
keep it the way it is. The link builder is highly dynamic and seems like
it would be tricky to make non-dynamic. Needs more thought.

Refs #3479