TableBuilder spec: Add a couple TODOs
It's not actually a collection of `workbenches` as we originally
thought, it's a collection of referentials. The name of the variable
should be changed to reflect that.

Also, in the real code, this value is not an array, but an
`ActiveRecord::Relation`, thus the `sortable_columns` method is able to
call `.model` on it. Since the test uses an array, that breaks. Need to
figure out what we want to do there. I don't like the dynamic-ness of
the way this is done now, but we need to figure out if it's possible to
factor that out in a pleasant way.

Refs #3479