TableBuilder spec: Stub `#model` on collection
`referentials` (known as `collection` in the `TableBuilder`) is supposed
to be an `ActiveRecord::Relation` instance. The `#sortable_columns`
method calls `collection.model` in order to dynamically build a table
column header. In order to get the test to the next stage, stub the
`#model` method. Not a huge fan. Ideally, the code wouldn't be so
dynamic, but maybe I'm wrong about that. Still need to think about this
for the refactoring.

Refs #3479