TableBuilder spec: Beautify actual HTML result
In order to allow us to easily compare the HTML result with the expected
value, beautify the HTML that comes out of the helper call.

Otherwise, all the HTML is on a single line, and it can't be compared
with RSpec's diff output (even if RSpec outputted the entire contents of
the single line, we wouldn't really be able to compare it).

Thanks to this post for describing an easy way to get HTML
beautification:
https://stackoverflow.com/questions/5622269/how-to-beautify-xml-code-in-rails-application

We need to run the expected HTML string through the beautifier so we can
get the formatting right, but otherwise this gives us a better base for
test diffing.

Refs #3479