TableBuilder spec: Use 'htmlbeautifier' gem for beautifying
Instead of `REXML::Document`, use `htmlbeautifier` to beautify our
actual HTML result. The trouble with `REXML` was that it forced single
quotes around attribute values. With 'htmlbeautifier', we get a good
result that we can realistically use for line-by-line diffs between
expected and actual. Much easier see what went wrong this way. And, in
my current case, see what I need to update update in the expected string
based on the factory-generated objects that I'm using.

Refs #3479