TableBuilder spec: Correct expected column header hrefs
Use the correct workbench ID in the expected links for column sorting.
Previously I had copied this from rendered HTML on my local dev server.
We should actually use the ID of the factory-generated workbench object.

Now this part of our tests passes, correctly matching the URLs.

Needed to copy over the `q` params that are created by the
`WorkbenchesController` in order for this to actually work with the
right query string parameters tacked on to the hrefs. Makes the test a
bit more complicated, but more closely matches the real page's
behaviour.

Hrm, I wonder if we should get rid of that param entirely and
modify the expected hrefs to removed them since that doesn't really
appear to be related to what we're actually testing. I'll consider it.

Refs #3479