TableBuilder spec: Add a test for sortable & non-sortable columns
Uses the table on the `/referentials/:id/companies` page. In that table,
all columns are sortable except for the "ID Codif" column.

The setup for the test is quite verbose and I don't like it.

I'll be using the test to change the sortable logic, removing the
hard-coded column header names checked inside the table builder in
favour of a flag that gets passed as part of the arguments to
`table_builder`.

Refs #3479