TableBuilder#build_column_header: Take Column instead of key & sortable
Now that we have a real `Column` type, it's much easier to just pass a
`Column` object to the method instead of these disconnected attributes.
Now we can get the column attributes from the object itself.

Remove the `#key_or_name` method as it's no longer used.

Also, I decided to reorder the `#build_column_header` arguments to make
`column` the first argument. Figured this made more sense as the method
is all about a column, and the collection model is this weird secondary
thing.

Refs #3479