TableBuilder#build_column_header: Pass `params` explicitly
I don't like that the method is using a global variable. It makes it
harder to read. Still don't like that we even use `params` in this
method. It would be a lot better if we could extricate it completely,
but I'll settle for this for now as there are other refactorings to get
to.

Refs #3479