workbenches/show.html.slim: Convert table_builder to new columns
Use our new column syntax to build the table, creating
`TableBuilderHelper::Column` objects for each column. This gets our
table rendering again with the new `table_builder_2` code.

Remove the old table builder code as we're going to switch entirely to
this one.

Get rid of the TODO about the =>s as it's no longer relevant.

Refs #3479