TableBuilder: Trial refactor of `#build_links`
Instead of the existing code that we have for `if policy...` etc. in
`#build_links`, create a few new methods that ostensibly do the same
thing using my `Link` class.

It should create a bunch of `Link`s that can be combined with those that
come from the `[TABLE-OBJECT].action_links` (like those in
`ReferentialDecorator#action_links`). The combination of these will then
be turned into `link_to`s by our existing code.

Renamed the `actions` argument to `#build_links` to `links` because they
come into `table_builder_2` as a `links` variable and I got confused by
the name not being the same.

The policy checking is now in its own method,
`#actions_after_policy_check`. Need to verify that the logic is right,
but otherwise it keeps things separate. Might also want to write special
methods for each of the conditions between the `||`s.

TODO: Move the new private methods to a class. Also test them.

Refs #3479