Move ModelDecorator to a new file; Remove duplicate PaginatingDecorator
Move `ModelDecorator` to a new file now that we know it works and that
we're going to use it. I had previously put these two classes in the
same file as `ReferentialDecorator` because I wanted to test that it
worked before making the code correct.

Also remove the `PaginatingDecorator` because apparently we already have
one, which I didn't realise. Use the existing one instead.

Refs #3479