WorkbenchesController: Remove `referential_decorator` import
This is no longer necessary. I had put it in because I was getting an
error, but it turns out that error was probably a result of me
redefining `PaginatingDecorator` inside "referential_decorator.rb". Now
that we're using the `PaginatingDecorator` that Jean-Paul created
originally, we no longer have a problem.

Refs #3479