TableBuilderHelper#thead: Remove unnecessary comment
I think this comment is no longer necessary. I had added it to the
original table builder code to help me understand it, but hopefully the
code is understandable enough. Also, we do the same thing in `#tbody`
and don't have a comment there, so I figured we could be consistent
about it.

Refs #3479