TableBuilder;Link: Remove default `:get` for HTTP method
Don't set links as `method: :get` by default. This shows up as:

    link_to _, _, method: :get

If we set the method to `nil` instead, we don't get a
`data-method="get"` attribute in our HTML output. Instead, we get no
`data-method` attribute, which is a lot better and cleaner.

I had originally used `:get` as a default because I wanted to generalise
links and certain ones we create need `:delete` or `:put` methods. I
figured, since we're always going to be passing `method:` to `link_to`,
we should have a sane default for that option. However, using `nil` is
even better as a default because then we don't get an extra attribute in
our HTML at all.

Refs #3479